pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "beefree-mcp-example"
version = "0.1.0"
description = "Showcase application for Beefree SDK MCP with PydanticAI"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    { name = "Beefree SDK", email = "dev@beefree.io" },
]
keywords = ["beefree", "pydantic-ai", "mcp", "email", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "pydantic-ai-slim[mcp,openai]>=1.0.8",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=13.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.scripts]
beefree-mcp-example = "main:main"

[tool.hatch.build.targets.wheel]
packages = ["."]
exclude = [
    "/.env*",
    "/.git*",
    "/static",
    "*.pyc",
    "__pycache__",
]

